Keep CSRF tokens valid for longer
Invalidating the CSRF token on each form submit is inconvenient: it
prevents users from using the tool in multiple tabs at once, because the
different tabs will fight over the last valid CSRF token. It seems to be
acceptable to keep the token valid for longer (MediaWiki’s CSRF tokens
are also valid for a long time); since this tool has no way to log in,
there’s no convenient place to clear the token at least once in a while
(like in other tools of mine), but that also means that the risk from
CSRF attacks should be low anyways, so I think it’s okay if the token
just stays valid forever (or until the session cookie expires).