Make selection keyboard-accessible
Previously, we hid the inputs using display: none, which removes them
from the tab order. Instead, use Bootstrap’s sr-only class, and add an
outline around the focused image (using the recent :focus-visible
pseudo-class, so that clicking the image does not activate it). To avoid
this outline overriding the existing “selected” indicator on an image,
introduce an extra element.