Increase outline size to cover whole image
We load the image with a maximum width and height of 250px, so if the
image is (nearly) square and actually takes up more than 200px in both
dimensions, then a 100px outline from each side isn’t enough to cover
it, leaving a small uncovered rectangle in the middle.

One example of this is [[File:2011 Rostelecom Cup - Hanyu-1.jpg]] [1] in
PagePile 25439 [2][3] (last image in the third row).

[1]: https://commons.wikimedia.org/wiki/File:2011_Rostelecom_Cup_-_Hanyu-1.jpg
[2]: https://pagepile.toolforge.org/api.php?action=get_data&id=25439&format=html
[3]: https://pagepile-visual-filter.toolforge.org/pagepile/25439/