Tell browser to load images lazily
Users were reporting HTTP 429 errors on large PagePiles, because the
browser is requesting too many images at once (basically T266155). To
alleviate this problem, tell the browser to lazily load the images only
as they scroll into view (or near the viewport).

For this to have any effect, we also need to add a width= and height= to
each <img>, otherwise all the images will be “visible” immediately and
nothing will be lazy-loaded. And since the width= and height= must have
the right aspect ratio (otherwise the image will be badly stretched),
this means we need to ask the API for the image sizes. That also gives
us proper thumbnail URLs, so we can bypass Special:FilePath; the API
additionally returns responsive URLs, so we might as well use that for a
srcset= on the <img>. (I tested that the API call itself doesn’t cause
the thumbnails to be generated yet, which is what we want.)

This slows down the loading of a large PagePile (because we now have to
make server-side API requests), but I think that’s an acceptable price.
It’s not as if such PagePiles would have instantly loaded client-side
before, the browser still needs to get all those images after all.

As part of the API loading, we also now don’t even generate markup for
PagePile entries that aren’t files (their imageinfo is None, which we
don’t render in the template). For now that should be fine, though we
also have the information available that would let the template show
non-files in a separate section, if needed.

Note that this doesn’t completely resolve the problem of 429 errors, so
I might add some JS code to gradually reload failed images later. But
for now I think this is better than nothing.