Automatically reload images on errors
This is intended for HTTP 429 Too Many Requests errors, but there’s no
way to get the error code, so it really applies to all errors.

I thought for a while about how to implement this in a responsible way.
The classic approach would be some kind of exponential backoff, but I
don’t think that’s ideal when the images are lazy-loaded: once the
current batch of in-flight images has fully loaded, there’s not much
reason to slow down the loading of the next batch, once the user has
scrolled down further. And we don’t want the delay to be specific to
each image anyways: if the user scrolled 200 images into view, causing
the browser to try loading 200 images at once, we shouldn’t send 160
retry requests as soon as the first 160 errors happen simultaneously.
For a while I thought we should reload images purely sequentially (e.g.
have one Promise for the current chain of reloads, and each error chains
its reload to the end of that), but that seemed a bit too slow, while at
the same time still potentially sending a busy loop of requests if the
server immediately responds with an error for a file for some reason.

Eventually, I came to the conclusion that what we want to limit is
*concurrency*. That is how the Thumbor limits in production are
implemented – up to 4 parallel requests with 50 more queued, with a
timeout of 4 seconds. We want to make sure that each reload starts at
least 100 milliseconds after the previous one – we may still have
several requests being served in parallel if the previous one didn’t
finish within 100 milliseconds, but we shouldn’t send more requests than
fit into the queue (though the queue may be partially emptied via
timeout instead of thumbnail completion). The minimum delay would be 80
milliseconds (4 seconds / 50), 100 milliseconds was just chosen to leave
some breathing room and also have a round number. Under this approach,
if the server immediately responds with an error for some URL, we only
repeat requests at a rate of at most 10 per second, not as fast as the
network allows.

I thought about adding some additional per-image delay on top of the
global 100-milliseconds-between-requests delay (i.e., this specific
image has been reloaded 5 times already, slow down a bit), but I don’t
think that would work well – it just means the image gets pushed into
the time frame that’s supposed to belong to the reloading of *other*
images. I think the only thing we can do per-image is to cut off the
reloads (e.g. stop reloading after 100 attempts), which would be more
useful for the “permanent error” case – but I’m not sure what a good
threshold would be that doesn’t interfere with the reloading we want for
large PagePiles. I might add something like this later, but it’s not
included in this commit.