Add possibility to open file description page
Open file description pages in a new tab on middle click, ctrl+click, or
(on Mac) command+click. (Regular click remains reserved for toggling the
checkbox.) This gives users a way to access higher-resolution versions
of the file.

Also, reword the <noscript> warning to be less specific, and advertise
generic “quality-of-life improvements” instead – which includes not only
the old selection buttons and the clicking feature added here, but also
the automatic reloading of images of the previous commits.

This feature might not be very accessible, but I’m not sure how to fix
that… I don’t think we should add proper links below each image, that
would clutter the interface too much IMHO.