Improve selected image styles
The previous outline-based solution worked fine in Firefox, but not so
in Chrome, where the outline can extend past the image and start to
overlap other images. We can fix this without changing the DOM at all,
by instead adding a solid background color to the span under the img,
and then making the img semi-transparent if checked. To make the span
occupy the full height of the img, make it an inline-block; this also
improves the outline for keyboard-focused inputs.

Also, change the selection color from a generic red to a Wikimedia blue;
I wasn’t sure if this made sense at first, but when I tried it out it
felt immediately natural for the selected images to turn blue, like they
would e.g. in a file browser.